moves Editor to a different file
